Refactor project screenshot image fetch
