+++ utils/shortcodes.js
@@ -1,4 +1,5 @@
 const Image = require("@11ty/eleventy-img");
+const path = require("path");
 
 module.exports = {
   codepen: (url, defaultTab = "result", height = 600, preview = false) => {
@@ -14,6 +15,35 @@ module.exports = {
     <span><a href="${url}">See the pen</a> (<a href="${user_profile}">@${username}</a>) on <a href="https://codepen.io">CodePen</a>.</span>
     </p>
     <script async src="https://cpwebassets.codepen.io/assets/embed/ei.js"></script>`;
+  },
+
+  codepenImage: async function (id) {
+    const url = `https://shots.codepen.io/hexagoncircle/pen/${id}-1280.jpg`;
+
+    let metadata = await Image(url, {
+      widths: [800],
+      outputDir: "_site/assets/images",
+      urlPath: "/assets/images",
+      cacheOptions: {
+        duration: "2w",
+        directory: ".cache",
+        removeUrlQueryParams: false,
+      },
+      filenameFormat: function (id, src, width, format, options) {
+        const extension = path.extname(src);
+        const name = path.basename(src, extension).split("-")[0];
+        return `codepen-${name}.${format}`;
+      },
+    });
+
+    let imageAttributes = {
+      alt: "",
+      loading: "lazy",
+      decoding: "async",
+      "data-is-loading": true,
+    };
+
+    return Image.generateHTML(metadata, imageAttributes);
   },
 
   image: async function (
